Change type of client ids to include URIs
Introduce new client id types that contains Uris so that replicas can
respond to a client based on information in a command they wish
to return the result of.

Note that at this point the schema hasn't been modified to include
this in messages, empty URIs are passed around at the moment.