Implement client response
Added the ability for clients to maintain Capnp servers and to receive
response messages.

At the moment as a placeholder the client simply prints out the result
rather than operating on the result.

Clients also do not block on sending a request - they should block until
a result is received.

Clients also execute indefintely (fst @@ Lwt.wait() line) so that they
don't terminate after sending all their requests and then crash the
replica trying to send a response. This should be removed once the
blocking behaviour described above is fixed.