Reverting post election partial update
This allowed the commit index to be reduced.
Eg suppose 5 nodes, leader has a partial partition with 3/4 followers.
Those followers vote for [a,b] before partition but don't receive the
commit
4th does receive commit of [a,b].
By safety the new leader does propose [a,b]
But from the 4th follower's perspective reverts to [a] is committed
which breaks execution.